requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
openai>=1.0.0
streamlit>=1.30.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
